trafo lv current calc update
For trafo vector groups with a gap in the zero seq. model (i.e. Dyn), the current calculation of the trafo lv side is special. It sums up the currents of the branch elements attached to it. Before, it was only considering elements that were attached with their FROM side. This led to wrong currents, when i.e. a line was attached the wrong way, but gave no warnig to the user.
Now, it also considers the TO side of elements, so lines can be attached either way.