Generation active power should be positive not negative
According to the pandapower signing system (https://pandapower.readthedocs.io/en/latest/about/units.html#conventions), the generation active power 'p_mw' should be according to the generator's point of view not the consumer's point of view and hence be positive, I fixed that in the text. It's important since this is a beginner's tutorial and the convention should be communicated clearly to avoid confusion.