Feature/dump_to_geojson (#1731)
* added dump_to_geojson function

* updated use of MultiLineString to LineString

MultiLineString is not required. It is intended to have disconnected lines as a single feature. This should not be the case for lines in pandapower.

* added convert_crs to convert geodata between coordinate reference systems
updated dump_to_geojson to be more performant
added possibility to only dump parts of the network

* added support for crs in geojson
fixed bug in convert_crs
convert_crs now adds crs reference to dataframe

* updated to support pandapipes directly

pandapipes therefore does not need to implement a variation of this

* fixed minor bug in dumping branches to geojson

* Update CHANGELOG.rst

minor changes

---------

Co-authored-by: hkoertge <hjalmar.koertge@iee.fraunhofer.de>
Co-authored-by: Jan Wiemer <32835953+jwiemer112@users.noreply.github.com>