Update io_utils.py
As std_types are stored as dict, when converting them to dataframe without specifying column dtypes, the dtypes are not properly inferred specially in case of trafo where mixed dtypes exist for different columns. So the dtype of all columns become object. when saving to sql dbs, this results in creation of tables with "varchar" type for all columns. Later when reading from these tables, a lot of warnings and bugs are generated.